# inspiracion/panda-etl/backend/app/docs/swagger.yaml
openapi: 3.0.0
info:
  title: MagicETL API
  description: |
    API para MagicETL - Plataforma modular de ETL con capacidades de IA.
    Esta API proporciona endpoints para gestionar proyectos, procesos ETL, transformaciones con IA y más.
  version: 1.0.0
  contact:
    name: Equipo de Desarrollo MagicETL
    email: support@magicetl.com
servers:
  - url: /api/panda
    description: Servidor principal de MagicETL
...
paths:
  /v1/projects:
    get:
      summary: Obtiene todos los proyectos
      tags: [Proyectos]
      responses:
        '200':
          description: Lista de proyectos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
    post:
      summary: Crea un nuevo proyecto
      tags: [Proyectos]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreate'
      responses:
        '201':
          description: Proyecto creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
  /v1/projects/{project_id}:
    get:
      summary: Obtiene un proyecto por su ID
      tags: [Proyectos]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Detalles del proyecto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          description: Proyecto no encontrado
    put:
      summary: Actualiza un proyecto
      tags: [Proyectos]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdate'
      responses:
        '200':
          description: Proyecto actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          description: Proyecto no encontrado
    delete:
      summary: Elimina un proyecto
      tags: [Proyectos]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Proyecto eliminado
        '404':
          description: Proyecto no encontrado
  /v1/projects/{project_id}/processes:
    get:
      summary: Obtiene todos los procesos de un proyecto
      tags: [Procesos]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [PENDING, IN_PROGRESS, COMPLETED, FAILED, STOPPED]
      responses:
        '200':
          description: Lista de procesos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Process'
    post:
      summary: Crea un nuevo proceso en un proyecto
      tags: [Procesos]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProcessCreate'
      responses:
        '201':
          description: Proceso creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Process'
...
components:
  schemas:
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProjectCreate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        description:
          type: string
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
    Process:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        status:
          type: string
          enum: [PENDING, IN_PROGRESS, COMPLETED, FAILED, STOPPED]
        project_id:
          type: integer
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        details:
          type: object
        message:
          type: string
        output:
          type: object
    ProcessCreate:
      type: object
      required:
        - name
        - type
      properties:
        name:
          type: string
        type:
          type: string
        details:
          type: object
...